+++ changelog.md
@@ -4,6 +4,11 @@ _(Note: This changelog focuses on the major changes between the different
 versions. Therefore, it may not contain all changes. Especially smaller fixes or
 improvements may be omitted.)_
 
+## Next Version (2026-08-??)
+
+* __[new feature]__
+  A Dutch translation is added.
+
 ## Version 0.2.1 (2026-08-07)
 
 * __[new feature]__